+++ pages/referer.php
@@ -93,10 +93,15 @@ echo $fragment->parse('core/page/section.php');
 
         $sql = rex_sql::factory();
 
+        // modify to include end date in period because SQL BETWEEN includes start and end date, but DatePeriod excludes end date
+        // without modification an additional day would be fetched from database
+        $end = clone $filter_date_helper->date_end;
+        $end->modify('+1 day'); 
+
         $period = new DatePeriod(
             $filter_date_helper->date_start,
             new DateInterval('P1D'),
-            $filter_date_helper->date_end
+            $end
         );
 
         foreach ($period as $value) {