+++ plugins/media/lib/stats_media_details.php
@@ -26,26 +26,6 @@ class stats_media_details
     }
 
 
-    // /**
-    //  *
-    //  *
-    //  * @return string
-    //  * @throws InvalidArgumentException
-    //  * @throws rex_exception
-    //  * @author Andreas Lenhardt
-    //  */
-    // public function get_list()
-    // {
-    //     $list = rex_list::factory('SELECT date, COUNT(date) as "count" FROM ' . rex::getTable('pagestats_dump') . ' WHERE url = "' . $this->url . '" GROUP BY date ORDER BY count DESC');
-    //     $list->setColumnLabel('date', 'Datum');
-    //     $list->setColumnLabel('count', 'Anzahl');
-    //     $list->setColumnSortable('date', $direction = 'desc');
-    //     $list->setColumnSortable('count', $direction = 'desc');
-    //     $list->setColumnParams('url', ['url' => '###url###']);
-
-    //     return $list->get();
-    // }
-
 
     /**
      *
@@ -77,10 +57,15 @@ class stats_media_details
     {
         $sql = rex_sql::factory();
 
+        // modify to include end date in period because SQL BETWEEN includes start and end date, but DatePeriod excludes end date
+        // without modification an additional day would be fetched from database
+        $end = clone $this->date_end;
+        $end->modify('+1 day');
+
         $period = new DatePeriod(
-            new DateTime($this->date_start->format('Y-m-d')),
+            $this->date_start,
             new DateInterval('P1D'),
-            new DateTime($this->date_end->format('Y-m-d'))
+            $end
         );
 
         foreach ($period as $value) {